Improved aiT profle import, compute edge criticalities, wca profile import, general code+documentation improvements
This patch adds support for computing edge criticalities (--compute-criticalities) [RTNS12], either using aiT or
platin/wca. It also adds import of wcet profiles (i.e., frequencies on the WCET path) and edge timings from both
aiT and platin/wca. It turns out import of aiT timings is much harder than expected. In particular it is not
always possible to map loop exit edges from aiT to LLVM, as loops are represented as functions in aiT. Finally,
this patch improves the code quality in general, with a more robust realization of PMLObjects and PMLLists.