platin: added pmlmod tool
Experimental tool that would handle limited changes to the PML analysis
information from the command line. The current use-case is removing (or
rewriting) information of unresolved (__any__) calls. A "PMLPath" (similar to
XPath for XML) is used to match certain PML elements (e.g. instructions) on
which a modification can be performed.