flowfacts: ConstantProgramPoint
A ConstantProgramPoint can be used within a flow fact of any type to add
a real constant offset to some flowfact. For example, to model that a
timer can activate an alarm every third activation:

   3 * AlarmActivations - C - __const(2) <= 0

The problem with just adding a real constant to the ILP was, that the
flow fact propagation eliminated it and moved the constant to the right
side, where it was related to the entry function (of the transformed
call-graph subtree).

TLDR: We can use flow facts with constants