Mangle filenames in qualify_machinefunction
The qnames for pml objects are based on a hierarchical string name, which
internally uses '/' as a component separator between function, block and
instruction identifiers. Using the path in an unmangled form will infere with
this structure, leading to broken parsing. Therefore, we normalize the path and
substitute slashes for '_'.