Implement 'type: callee' modelfacts
Implement support for type: callee annotations. Currently, the preliminary
expression format (to be finalizes) is:
  [globalfunction, path/to/defining/module.c:static_function]

Patmos-llvm currently namemangles functions with internal linkage (see
PMLExport.cpp for details) as follows:
   modulename =~ s/[^0-9A-Za-z]/_/g
   modulename + '_' + functionname

In most cases, modulename is the path to the .c source file the compiler sees.
For the example above, the mangled name is as follows:
   path_to_defining_module_c_static_function

Obviously, the namemangling-implementations of platin and patmos-clang have to
be kept in sync