Preliminary ILPVisualisation implementation
This is the initial draft for an ILP visualizer. This will be refactored to a
proper visualizer in tools/visualize.rb in the future.

It creates to files (currently in pwd), ilp.svg which is a dot-generated
version of the ILP and constraints.json which can be used with the
html based Viewer in ./ilpviewer