WCA/ILP-Visualisation: export vis to hash if requested
Up to now, the files (svg visualisation as svg, constraints info and
sourcehints) were only emitted to the outdir. In preparation for the html-based
interactive visualisation, we need the possibility to extract this information
programmatically and directly via a hash.

This is kind of a hack (as we use the options-struct to pass an result var by
reference), but kindof works and is the best we can get.