interact.rb: added an visualise command
Invocation:
 visualize ilp <entrypoint>

This starts the ilp visualisation server with an interactive debugging session
for the ilp starting from wcet-analysis entrypoint `entrypoint`

By default, the Server is started at port "2142" and binds on localhost only,
which can be changed via the --server-bind-addr or --server-port parameters
respectively.
A source-directory is required, it can be set by the "--source-path" parameter.