Peaches: Bugfix: eliminate parsing ambiguity
There is no clear distinction between arith_exprs and con_exprs that can be
decided upon parsing. The former, artificial distinction was therefore unsound.
It is the task of the typechecker to assert correct typeinformation.
Therefore, the differentiation in parsing is removed.

The differentiation is now enforced upon evaluation (ducktyping), but should be
fixed once HM-style inference is implemented.