Peaches: Fix handling of curried functions as parameters
Peaches failed to correctly bind curried functions when those were
passed as higher order function arguments.
This commit implements proper currying via ASTScopes. An ASTScope is an
ASTExpr which binds the already supplied values when the call-context is
reentered.
It can be used to implement `let x = 4, y=6 in f x y` style constructs
in the future.