Fix names for lpsolve constrants.
The unescaped '-' in the replacement regex leads to the replacement of every
character in the range from ')' to '>'. In particular, this leads to the
replacement of numbers. As numbers are not forbidden in lpsolve constraint
names, this commit fixes the regex, leading to more readable constraint names in
the lpsolve files.

Additionally, the characters '@' and '/' are removed from the regex, as the
lpsolve documentation does not name them as forbidden.