transform: change order of variable elimination
The order in which constraints are considered during variable
elimination matters, as the elimination of contraints by, for example,
infeasibility annotations, which are of the form 'v = 0', may lead to
loss of information. This patch sorts the known constraints additionally
by the number of unaffected variables, considering constraints with more
unaffected variables earlier. This reduces the likelihood of information
loss.